Ready for merge
git-svn-id: https://svn.atnf.csiro.au/difx@4407 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ mpifxcorr/branches/fakedatastream/src/configuration.cpp
@@ -1951,7 +1951,7 @@ bool Configuration::populateResultLengths()
 void Configuration::setDatastreamMuxInfo(int datastreamindex, string muxinfo)
 {
   int threadindices[500];
-  int at = muxinfo.find_first_of(":");
+  size_t at = muxinfo.find_first_of(":");
   datastreamtable[datastreamindex].nummuxthreads = 0;
 
   while(at != string::npos) {
